feat!(rust,python): Rust APIに`UserDictWordPriority`型を導入

+++ crates/voicevox_core_c_api/include/voicevox_core.h
@@ -508,7 +508,7 @@ typedef struct VoicevoxUserDictWord {
   /**
    * 優先度
    */
-  uint32_t priority;
+  uint8_t priority;
 } VoicevoxUserDictWord;
 
 #ifdef __cplusplus
